requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
langchain-openai==0.3.28
pytest==8.4.1